Extract Tmuxinator::Doctor utility class
Extracts methods from Tmuxinator::Config that were only invoked as part
of the `tmuxinator doctor` command.

https://github.com/tmuxinator/tmuxinator/issues/457
