+++ lib/tmuxinator.rb
@@ -11,6 +11,7 @@ require "tmuxinator/deprecations"
 require "tmuxinator/wemux_support"
 require "tmuxinator/cli"
 require "tmuxinator/config"
+require "tmuxinator/doctor"
 require "tmuxinator/hooks"
 require "tmuxinator/hooks/project"
 require "tmuxinator/pane"
